% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_ejemplos.R
\name{descargar_ejemplo}
\alias{descargar_ejemplo}
\title{Descargar ejemplos al directorio de trabajo}
\usage{
descargar_ejemplo(nombre_archivo = NULL)
}
\arguments{
\item{nombre_archivo}{Nombre del archivo que se quiere
descargar al directorio de trabajo (incluyendo su extensión).
Si no se indica el nombre de un archivo, se imprimirá en la
consola la lista de archivos disponibles en ese directorio.}
}
\description{
El paquete liftún incluye archivos en diversos formatos
para practicar la limpieza y transformación de datos.
Esta función permite descargar esos archivos al
directorio de trabajo para poder utilizar las
funciones de importación de datos que correspondan.
}
\examples{
descargar_ejemplo("visitas_snaspe.xlsx")
}
